[DDW-738] Move toggle component to discreet mode feature folder
