+++ source/renderer/app/features/discreet-mode/ui/discreet-toggle/DiscreetToggle.scss
