+++ storybook/stories/nodes/environment/TopBarEnvironment.stories.js
@@ -14,7 +14,7 @@ import WalletTestEnvironmentLabel from '../../../../source/renderer/app/componen
 import { formattedWalletAmount } from '../../../../source/renderer/app/utils/formatters';
 import menuIconClosed from '../../../../source/renderer/app/assets/images/menu-ic.inline.svg';
 import NewsFeedIcon from '../../../../source/renderer/app/components/widgets/NewsFeedIcon';
-import DiscreetToggle from '../../../../source/renderer/app/components/widgets/discreet-mode/DiscreetToggle';
+import DiscreetToggle from '../../../../source/renderer/app/features/discreet-mode/ui/discreet-toggle/DiscreetToggle';
 import { isShelleyTestnetTheme } from '../../_support/utils';
 
 const topBarTestEnv = (currentTheme) => (
